+++ source/Irrlicht/CGUIEditBox.cpp
@@ -446,7 +446,7 @@ bool CGUIEditBox::processKey(const SEvent& event)
 						newMarkBegin = CursorPos;
 
 					newMarkEnd = pos + 1;
-					if(!newMarkEnd || newMarkEnd > Text.size()) {
+					if(!newMarkEnd || newMarkEnd > (s32)Text.size()) {
 						newMarkEnd = Text.size();
 					}
 				}
@@ -455,7 +455,7 @@ bool CGUIEditBox::processKey(const SEvent& event)
 				newMarkEnd = 0;
 			}
 
-			if(Text.size() > pos) CursorPos = pos + 1;
+			if((s32)Text.size() > pos) CursorPos = pos + 1;
 			else CursorPos = Text.size();
 			BlinkStartTime = os::Timer::getTime();
 			break;
